start using usize more as the conversion is annoying, and storing >4gb files on 32-bit machines is not that important for now.

+++ binc/src/comments.rs
@@ -1,12 +1,15 @@
-
 #[derive(Debug, Default)]
 pub(crate) struct Comments {
     pub comments: Vec<Comment>,
 }
 
 impl Comments {
-    pub(crate) fn add_comment(&mut self, comment: &str, author: &str, response_to: u64) {
-        let comment = Comment { author: author.to_string(), text: comment.to_string(), response_to: Some(response_to as usize) };
+    pub(crate) fn add_comment(&mut self, comment: &str, author: &str, response_to: usize) {
+        let comment = Comment {
+            author: author.to_string(),
+            text: comment.to_string(),
+            response_to: Some(response_to),
+        };
         self.comments.push(comment);
     }
 }
